Compiler architecture update Design Flow Manager update  Design flow manager has been updated to avoid use of boost::filtered_graph.  Design flow step signature is now an integer instead of a string to improve performance.  Design flow manager profiling has been improved; CSV file generation has been added. Tree manager update  Tree manager is now handling tree_node instances by default instead of tree_reindex to avoid multiple indirection.  Tree reindex remove has been implemented to remove all tree_reindex instances after Bambu IR parsing is completed.  GET_NODE, GET_INDEX_NODE have been removed. Abseil update to version 20240116.2 included as submodule Parameter class template refactoring to improve string to type casting performance General code refactoring to reduce the usage of dynamic_cast Removed unused tree node memory_tag Pragma analysis step refactoring Bit value opt step refactoring Other minor fixes
Minor refactoring

Hide DesignFlowStepPrioritySet implementation

+++ src/EucalyptusParameter.cpp
@@ -228,11 +228,11 @@ int EucalyptusParameter::Exec()
          {
             if(std::string(optarg) == "V")
             {
-               setOption(OPT_writer_language, static_cast<int>(HDLWriter_Language::VERILOG));
+               setOption(OPT_writer_language, HDLWriter_Language::VERILOG);
             }
             else if(std::string(optarg) == "H")
             {
-               setOption(OPT_writer_language, static_cast<int>(HDLWriter_Language::VHDL));
+               setOption(OPT_writer_language, HDLWriter_Language::VHDL);
             }
             else
             {
@@ -642,7 +642,7 @@ void EucalyptusParameter::SetDefaults()
    setOption(OPT_parallel_backend, false);
 
    /// backend HDL
-   setOption(OPT_writer_language, static_cast<int>(HDLWriter_Language::VERILOG));
+   setOption(OPT_writer_language, HDLWriter_Language::VERILOG);
    setOption(OPT_reset_type, "no");
    setOption(OPT_reg_init_value, false);
